requirements.txt
psycopg2-binary==2.9.9
requests==2.32.3
pendulum==3.0.0
pytest==8.3.3
pytest-mock==3.14.0
mypy==1.11.2
ruff==0.6.9
types-requests==2.32.0.20240914